Support building android-wireguard with GN
This commit adds a BUILD.gn to the repository which enables us to build
directly via Chromium's build system. We already use this repository
within Brave but via a hand-built AAR that gets extracted. With this
commit, we have first-class support for building the repo from source
during the brave build.

To address the problem that Chromium cannot build Java sources that use
the `record` type, the singular instance of a record has been converted
to the standard `class` format.